requirements.txt.txt
python-telegram-bot==20.3
requests==2.31.0
web3==6.10.0